% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ttest.R
\name{calc_ttest}
\alias{calc_ttest}
\title{Perform a Two-Sample T-Test}
\usage{
calc_ttest(data, var, group_var)
}
\arguments{
\item{data}{Data frame containing the two groups}

\item{var}{Name of the variable to test}

\item{group_var}{Name of the grouping variable (must have 2 levels)}
}
\value{
Tibble with the p-value and 95\% confidence interval
}
\description{
This function performs a two-sample t-test to compare means between two groups.
}
