% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_analysis.R
\name{corr_analysis}
\alias{corr_analysis}
\title{Calculate Pearson Correlation Coefficient}
\usage{
corr_analysis(data, x_var, y_var)
}
\arguments{
\item{data}{Data frame containing the variables}

\item{x_var}{Name of the first variable}

\item{y_var}{Name of the second variable}
}
\value{
Tibble with the correlation coefficient and p-value
}
\description{
This function calculates the Pearson correlation between two variables.
}
